🚚 Prepare move to MQT organization (#127)
This PR prepares the move to the MQT organization. Furthermore, it

- updates MQT Core,
- updates the workflows,
- streamlines the Python bindings,
- streamlines the README and documentation,
- sets up templating,
- drops support for Python 3.9,
- drops support for x86 macOS systems, and
- further aligns the repository with other MQT repositories.

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

+++ test/python/core/test_lanes.py
@@ -1,3 +1,11 @@
+# Copyright (c) 2023 - 2025 Chair for Design Automation, TUM
+# Copyright (c) 2025 Munich Quantum Software Company GmbH
+# All rights reserved.
+#
+# SPDX-License-Identifier: MIT
+#
+# Licensed under the MIT License
+
 from __future__ import annotations
 
 import typing
@@ -24,7 +32,11 @@ class TestLanes(TestCase):
             CEx(self.circuit, "CEx", [0, 1], None, [self.circuit.dimensions[i] for i in [0, 1]]),
             R(self.circuit, "R", 1, [0, 1, np.pi, np.pi / 2], self.circuit.dimensions[1]),
             CustomMulti(
-                self.circuit, "CUm", [0, 1, 2], np.identity(18), [self.circuit.dimensions[i] for i in [0, 1, 2]]
+                self.circuit,
+                "CUm",
+                [0, 1, 2],
+                np.identity(18, dtype=np.complex128),
+                [self.circuit.dimensions[i] for i in [0, 1, 2]],
             ),
         ]
         self.circuit.instructions = gates
@@ -43,7 +55,11 @@ class TestLanes(TestCase):
             CEx(self.circuit, "CEx", [0, 1], None, [self.circuit.dimensions[i] for i in [0, 1]]),
             R(self.circuit, "R", 1, [0, 1, np.pi, np.pi / 2], self.circuit.dimensions[1]),
             CustomMulti(
-                self.circuit, "CUm", [0, 1, 2], np.identity(18), [self.circuit.dimensions[i] for i in [0, 1, 2]]
+                self.circuit,
+                "CUm",
+                [0, 1, 2],
+                np.identity(18, dtype=np.complex128),
+                [self.circuit.dimensions[i] for i in [0, 1, 2]],
             ),
         ]
         self.circuit.instructions = gates
@@ -58,7 +74,11 @@ class TestLanes(TestCase):
             CEx(self.circuit, "CEx", [0, 1], None, [self.circuit.dimensions[i] for i in [0, 1]]),
             R(self.circuit, "R", 1, [0, 1, np.pi, np.pi / 2], self.circuit.dimensions[1]),
             CustomMulti(
-                self.circuit, "CUm", [0, 1, 2], np.identity(18), [self.circuit.dimensions[i] for i in [0, 1, 2]]
+                self.circuit,
+                "CUm",
+                [0, 1, 2],
+                np.identity(18, dtype=np.complex128),
+                [self.circuit.dimensions[i] for i in [0, 1, 2]],
             ),
         ]
         self.circuit.instructions = gates
@@ -78,7 +98,11 @@ class TestLanes(TestCase):
             R(self.circuit, "R", 0, [0, 1, np.pi, np.pi / 2], self.circuit.dimensions[0]),
             R(self.circuit, "R", 0, [0, 1, np.pi, np.pi / 2], self.circuit.dimensions[0]),
             CustomMulti(
-                self.circuit, "CUm", [0, 1, 2], np.identity(18), [self.circuit.dimensions[i] for i in [0, 1, 2]]
+                self.circuit,
+                "CUm",
+                [0, 1, 2],
+                np.identity(18, dtype=np.complex128),
+                [self.circuit.dimensions[i] for i in [0, 1, 2]],
             ),
             R(self.circuit, "R", 2, [0, 1, np.pi, np.pi / 2], self.circuit.dimensions[2]),
             CEx(self.circuit, "CEx", [0, 1], None, [self.circuit.dimensions[i] for i in [0, 1]]),
